requirements.txt
transformers==4.48.0
flask==3.1.0
torch==2.5.1
sacremoses==0.1.1
sentencepiece==0.2.0
pillow==11.1.0